+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-### 3.14.0(45)
+### [3.14.0(45)](https://github.com/VakhoKontridze/VCore/releases/tag/3.14.0) — *2022 07 12*
 
 Views and ViewControllers